move: 종목 지정가와 지정가 알림 분리
